# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/message": {
      "post": {
        "operationId": "MessageController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建消息（支持全员、部分、个人通知）",
        "tags": [
          "信息管理"
        ]
      },
      "get": {
        "operationId": "MessageController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取当前用户所有消息（含全员通知）",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/conversations": {
      "get": {
        "operationId": "MessageController_getPrivateConversations",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "游标字符串",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取当前用户的私信会话列表",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/conversations/{userId}/messages": {
      "get": {
        "operationId": "MessageController_getPrivateConversation",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "会话对方用户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "游标字符串",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取与指定用户的私信记录",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/read-batch": {
      "post": {
        "operationId": "MessageController_markPrivateMessagesRead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchReadPrivateMessagesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量标记私信已读",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/{userId}": {
      "post": {
        "operationId": "MessageController_sendPrivateMessage",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "接收者用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPrivateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发送私信",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/recall/{id}": {
      "post": {
        "operationId": "MessageController_recallPrivateMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "私信ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallPrivateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "撤回私信",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/block/{userId}": {
      "post": {
        "operationId": "MessageController_blockPrivateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "目标用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "拉黑私信对象",
        "tags": [
          "信息管理"
        ]
      },
      "delete": {
        "operationId": "MessageController_unblockPrivateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "目标用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "取消拉黑私信对象",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/private/blocks": {
      "get": {
        "operationId": "MessageController_getBlockedUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取拉黑列表",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/search": {
      "get": {
        "operationId": "MessageController_search",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "消息类型",
            "schema": {
              "enum": [
                "private",
                "system",
                "notification"
              ],
              "type": "string"
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "description": "是否已读",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isBroadcast",
            "required": false,
            "in": "query",
            "description": "是否为广播消息",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "搜索关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senderId",
            "required": false,
            "in": "query",
            "description": "发送者ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "receiverId",
            "required": false,
            "in": "query",
            "description": "接收者ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "高级查询消息",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/{id}": {
      "get": {
        "operationId": "MessageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "消息ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取单条消息",
        "tags": [
          "信息管理"
        ]
      },
      "patch": {
        "operationId": "MessageController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "消息ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新消息内容",
        "tags": [
          "信息管理"
        ]
      },
      "delete": {
        "operationId": "MessageController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "消息ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除消息",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/{id}/read": {
      "post": {
        "operationId": "MessageController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "消息ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "标记消息为已读",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/read-all": {
      "post": {
        "operationId": "MessageController_markAllAsRead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkAllReadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "标记所有消息为已读",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/batch": {
      "post": {
        "operationId": "MessageController_batchOperation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量操作消息（标记已读/删除）",
        "tags": [
          "信息管理"
        ]
      }
    },
    "/message/unread/count": {
      "get": {
        "operationId": "MessageController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取未读消息数量",
        "tags": [
          "信息管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "PicArt API",
    "description": "PicArt 图片社区 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "object",
            "description": "发送者ID",
            "example": 1
          },
          "receiverId": {
            "type": "number",
            "description": "接收者ID（单发）",
            "example": 2
          },
          "receiverIds": {
            "description": "接收者ID列表（批量）",
            "example": [
              2,
              3,
              4
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "content": {
            "type": "string",
            "description": "消息内容",
            "example": "这是一条消息"
          },
          "title": {
            "type": "string",
            "description": "消息标题",
            "example": "消息标题"
          },
          "type": {
            "type": "string",
            "description": "消息类型",
            "enum": [
              "private",
              "system",
              "notification"
            ],
            "default": "private"
          },
          "isBroadcast": {
            "type": "boolean",
            "description": "是否为广播消息",
            "default": false
          },
          "metadata": {
            "type": "object",
            "description": "消息元数据"
          }
        },
        "required": [
          "content",
          "type",
          "isBroadcast"
        ]
      },
      "BatchReadPrivateMessagesDto": {
        "type": "object",
        "properties": {
          "messageIds": {
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "messageIds"
        ]
      },
      "SendPrivateMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "文本内容"
          },
          "messageKind": {
            "type": "string",
            "description": "消息类型",
            "enum": [
              "text",
              "image",
              "file",
              "card"
            ],
            "default": "text"
          },
          "payload": {
            "type": "object",
            "description": "结构化负载"
          }
        },
        "required": [
          "messageKind"
        ]
      },
      "RecallPrivateMessageDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "撤回原因"
          }
        }
      },
      "UpdateMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "消息内容",
            "example": "更新的消息内容"
          },
          "title": {
            "type": "string",
            "description": "消息标题",
            "example": "更新的消息标题"
          },
          "isRead": {
            "type": "boolean",
            "description": "是否已读"
          },
          "type": {
            "type": "string",
            "description": "消息类型",
            "enum": [
              "private",
              "system",
              "notification"
            ]
          },
          "isBroadcast": {
            "type": "boolean",
            "description": "是否为广播消息"
          },
          "metadata": {
            "type": "object",
            "description": "消息元数据"
          }
        }
      },
      "MarkAllReadDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "消息类型",
            "enum": [
              "private",
              "system",
              "notification"
            ]
          },
          "isBroadcast": {
            "type": "boolean",
            "description": "是否为广播消息"
          }
        }
      },
      "BatchMessageDto": {
        "type": "object",
        "properties": {
          "messageIds": {
            "description": "消息ID列表",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "action": {
            "type": "string",
            "description": "操作类型",
            "enum": [
              "read",
              "delete"
            ],
            "example": "read"
          }
        },
        "required": [
          "messageIds",
          "action"
        ]
      },
...
    }
  }
}